Add files via upload
Udemy dataset for the year 2011-2017 helps to analyze & visualize which course is highly viewed or purchased, which courses udemy should focus on, when udemy should publish their new courses. The project aims to convey the analysis of different courses is paid or not from the data from the internet. The python libraries used in the exploratory data analysis include NumPy, Pandas, Matplotlib, Seaborn.